fix(anchor-link-select): refresh anchor list on every render
